fix conflicrs after pulling from dev, also added second pswd reveal icon in signup form
