+++ src/screens/Auth/ExpiredLink/ExpiredLink.scss
@@ -1,5 +1,5 @@
-@import '../../scss/variables';
-@import '../../scss/mixins';
+@import 'scss/variables';
+@import 'scss/mixins';
 
 .expired-page {
   align-items: center;
