+++ src/screens/Auth/index.ts
@@ -1,3 +1,5 @@
+export * from './EmailVerify/EmailVerify'
+export * from './ExpiredLink/ExpiredLink'
 export * from './RegisterUserInfo/RegisterUserInfo'
 export * from './SignIn/SignIn'
 export * from './SignUp/SignUp'
